Move getOAuthServerUrl into oauth-utils
It is OAuth-specific by name and purpose and only used by OAuth code, so it
belongs in the shared OAuth utilities module rather than the generic utils.
Pure move (function + tests relocated); no behaviour change.
